Skip interactive_graphics test on Python 3.4

+++ wntr/tests/test_graphics.py
@@ -1,6 +1,7 @@
 from nose.tools import *
+from nose import SkipTest
 from os.path import abspath, dirname, join, isfile
-import os
+import os, sys
 import wntr
 import matplotlib.pylab as plt
 
@@ -24,6 +25,8 @@ def test_plot_network1():
     assert_true(isfile(filename))
 
 def test_plot_interactive_network1():
+    if (sys.version_info.major == 3) and (sys.version_info.major == 4):
+        raise SkipTest # skip if python version = 3.4
     filename = abspath(join(testdir, 'plot_interactive_network1.html'))
     if isfile(filename):
         os.remove(filename)
